refactor: update math functions to use usize instead of u32, and delete unused string.zig
- Change alignUp, alignDown, and isAligned to use usize for addresses and alignment values
  for better portability across 32-bit and 64-bit systems
- Delete src/lib/string.zig (dead code: stringCopy and stringCompare unused)

+++ src/lib/math.zig
@@ -1,11 +1,11 @@
-pub fn alignUp(addr: u32, alignment: u32) u32 {
+pub fn alignUp(addr: usize, alignment: usize) usize {
     return (addr + alignment - 1) & ~(alignment - 1);
 }
 
-pub fn alignDown(addr: u32, alignment: u32) u32 {
+pub fn alignDown(addr: usize, alignment: usize) usize {
     return addr & ~(alignment - 1);
 }
 
-pub fn isAligned(addr: u32, alignment: u32) bool {
+pub fn isAligned(addr: usize, alignment: usize) bool {
     return (addr & (alignment - 1)) == 0;
 }
